backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
pydantic
aiofiles
python-dotenv
